Save prediction to db

+++ db/Create_Predictions_Table.sql
@@ -7,7 +7,7 @@ CREATE TABLE dbo.Predictions
 (
   Id INT NOT NULL IDENTITY PRIMARY KEY,
   Prediction VARCHAR(50) NOT NULL,
-  Difficulty NUMERIC(4,3) NOT NULL,
+  Difficulty tinyint NOT NULL,
   ImageUri VARCHAR(250) NOT NULL,
   CreatedBy VARCHAR(50) NOT NULL
 );
